working through client transforms
